予測＋スケジュールがGPVのデータを使って、一挙に動作する #14予測アルゴリズムの統合 #4
