+++ main.py
@@ -10,7 +10,7 @@ subprocess.run(['python', 'weather_data.py'])
 #PV出力予測
 subprocess.run(['python', 'pv_predict.py'])
 
-#price_forecast.pyを実行する
+##price_forecast.pyを実行する
 subprocess.run(['python', 'price_predict.py'])
 
 # ESS_control.pyを実行する
